+++ benchmark/src/main/java/feign/benchmark/DecoderIteratorsBenchmark.java
@@ -1,14 +1,26 @@
-package feign.stream;
+/**
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package feign.benchmark;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import feign.Response;
+import feign.Util;
 import feign.codec.Decoder;
 import feign.jackson.JacksonDecoder;
 import feign.jackson.JacksonIteratorDecoder;
+import feign.stream.StreamDecoder;
 import org.openjdk.jmh.annotations.*;
-import org.openjdk.jmh.runner.Runner;
-import org.openjdk.jmh.runner.options.Options;
-import org.openjdk.jmh.runner.options.OptionsBuilder;
 
 import java.lang.reflect.Type;
 import java.util.Collection;
@@ -18,13 +30,17 @@ import java.util.List;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Stream;
 
-import static feign.Util.UTF_8;
-
+/**
+ * This test shows up how fast different json array response processing implementations are.
+ */
 @State(Scope.Thread)
 public class DecoderIteratorsBenchmark {
 
   @Param({"list", "iterator", "stream"})
   private String api;
+
+  @Param({"10", "100"})
+  private String size;
 
   private Response response;
 
@@ -64,7 +80,7 @@ public class DecoderIteratorsBenchmark {
         .status(200)
         .reason("OK")
         .headers(Collections.emptyMap())
-        .body(carsJson(10), UTF_8)
+        .body(carsJson(Integer.valueOf(size)), Util.UTF_8)
         .build();
   }
 
@@ -104,13 +120,5 @@ public class DecoderIteratorsBenchmark {
   static class Car {
     public String name;
     public String manufacturer;
-  }
-
-  public static void main(String[] args) throws Exception {
-    Options opt = new OptionsBuilder()
-        .include(DecoderIteratorsBenchmark.class.getSimpleName())
-        .build();
-
-    new Runner(opt).run();
   }
 }
